Mongodb/documents

+++ src/main/java/br/org/institutobushido/services/turma/TurmaService.java
@@ -2,17 +2,26 @@ package br.org.institutobushido.services.turma;
 
 import java.util.List;
 
+import br.org.institutobushido.resources.exceptions.LimitQuantityException;
+
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.aggregation.Aggregation;
 import org.springframework.data.mongodb.core.aggregation.AggregationOperation;
 import org.springframework.data.mongodb.core.query.Criteria;
+import org.springframework.data.mongodb.core.query.Query;
+import org.springframework.data.mongodb.core.query.Update;
 import org.springframework.stereotype.Service;
 
-import br.org.institutobushido.controllers.dtos.turma.TurmaAlunoResponse;
+import br.org.institutobushido.controllers.dtos.turma.TurmaAlunoDTOResponse;
 import br.org.institutobushido.controllers.dtos.turma.TurmaDTORequest;
 import br.org.institutobushido.controllers.dtos.turma.TurmaDTOResponse;
 import br.org.institutobushido.mappers.turma.TurmaMapper;
+import br.org.institutobushido.models.admin.Admin;
+import br.org.institutobushido.models.admin.turmas.TurmaResponsavel;
+import br.org.institutobushido.models.aluno.Aluno;
 import br.org.institutobushido.models.turma.Turma;
+import br.org.institutobushido.repositories.AdminRepositorio;
 import br.org.institutobushido.repositories.TurmaRepositorio;
 import br.org.institutobushido.resources.exceptions.AlreadyRegisteredException;
 import br.org.institutobushido.resources.exceptions.EntityNotFoundException;
@@ -21,32 +30,50 @@ import br.org.institutobushido.resources.exceptions.EntityNotFoundException;
 public class TurmaService implements TurmaServiceInterface {
 
     private final TurmaRepositorio turmaRepositorio;
-    private MongoTemplate mongoTemplate;
+    private final MongoTemplate mongoTemplate;
+    private final AdminRepositorio adminRepositorio;
 
-    public TurmaService(TurmaRepositorio turmaRepositorio, MongoTemplate mongoTemplate) {
+    public TurmaService(TurmaRepositorio turmaRepositorio, MongoTemplate mongoTemplate,
+            AdminRepositorio adminRepositorio) {
         this.mongoTemplate = mongoTemplate;
         this.turmaRepositorio = turmaRepositorio;
+        this.adminRepositorio = adminRepositorio;
     }
 
     @Override
-    public String criarNovaTurma(TurmaDTORequest turma) {
+    public String criarNovaTurma(String emailAdmin, TurmaDTORequest turma) {
         boolean turmaExiste = this.verificaSeTurmaExiste(turma.nome());
 
         if (turmaExiste) {
-            throw new AlreadyRegisteredException("Turma com esse nome ja está registrada");
+            throw new AlreadyRegisteredException("Turma " + turma.nome() + " ja está registrada.");
         }
 
         Turma novaTurma = TurmaMapper.mapToTurma(turma);
+
+        Admin admin = this.vinculrTurmaAoAdmin(emailAdmin,
+                new TurmaResponsavel(novaTurma.getEndereco(), novaTurma.getNome()));
+
+        novaTurma.setTutor(admin.getNome());
+
         this.turmaRepositorio.save(novaTurma);
-        return "Turma criada " + turma.nome() + " com sucesso!";
+        return "Turma " + turma.nome() + " foi criada com sucesso!";
     }
 
     @Override
-    public String deletarTurma(String nomeTurma) {
+    public String deletarTurma(String emailAdmin, String nomeTurma) {
+
         boolean turmaExiste = this.verificaSeTurmaExiste(nomeTurma);
+
         if (!turmaExiste) {
             throw new EntityNotFoundException("Turma com esse nome não existe");
         }
+
+        if (this.vefrificarSeExistemAlunosAtivosNaTurma(nomeTurma)) {
+            throw new LimitQuantityException("Turma com alunos não pode ser deletada");
+        }
+
+        this.desvicularTurmaDoAdmin(emailAdmin, nomeTurma);
+
         this.turmaRepositorio.deleteByNome(nomeTurma);
         return "Turma " + nomeTurma + " deletada com sucesso";
     }
@@ -57,13 +84,14 @@ public class TurmaService implements TurmaServiceInterface {
         return TurmaMapper.mapToListTurmaDTOResponse(turmas);
     }
 
+    @Cacheable(value = "turma", key = "#nomeTurma")
     @Override
     public TurmaDTOResponse buscarTurmaPorNome(String nomeTurma) {
         return TurmaMapper.mapToTurmaDTOResponse(this.encontrarTurmaPeloNome(nomeTurma));
     }
 
     @Override
-    public List<TurmaAlunoResponse> listarAlunosDaTurma(String nomeTurma) {
+    public List<TurmaAlunoDTOResponse> listarAlunosDaTurma(String nomeTurma) {
         AggregationOperation match = Aggregation.match(Criteria.where("nome").is(nomeTurma));
 
         AggregationOperation lookup = Aggregation.lookup("alunos", "nome", "turma", "alunos_turma");
@@ -78,16 +106,46 @@ public class TurmaService implements TurmaServiceInterface {
                 .andExclude("_id");
 
         Aggregation aggregation = Aggregation.newAggregation(match, lookup, unwind, project);
-        var result = mongoTemplate.aggregate(aggregation, "turmas", TurmaAlunoResponse.class);
-        return result.getMappedResults();
+        return mongoTemplate.aggregate(aggregation, "turmas", TurmaAlunoDTOResponse.class).getMappedResults();
     }
 
     private boolean verificaSeTurmaExiste(String nomeTurma) {
         return this.turmaRepositorio.findByNome(nomeTurma).isPresent();
     }
 
+    private Admin vinculrTurmaAoAdmin(String emailAdmin, TurmaResponsavel turma) {
+        Admin admin = this.encontrarAdminPeloEmail(emailAdmin);
+        Query query = new Query();
+        query.addCriteria(Criteria.where("email").is(emailAdmin));
+        Update update = new Update();
+        update.push("turmas", admin.adicionarTurma(turma));
+        this.mongoTemplate.updateFirst(query, update, Admin.class);
+        return admin;
+    }
+
+    private void desvicularTurmaDoAdmin(String emailAdmin, String nomeTurma) {
+        Admin admin = this.encontrarAdminPeloEmail(emailAdmin);
+        String turma = admin.removerTurma(nomeTurma);
+        Query query = new Query();
+        query.addCriteria(Criteria.where("email").is(emailAdmin));
+        Update update = new Update();
+        update.pull("turmas", Query.query(Criteria.where("nome").is(turma)));
+        this.mongoTemplate.updateFirst(query, update, Admin.class);
+    }
+
+    private boolean vefrificarSeExistemAlunosAtivosNaTurma(String nomeTurma) {
+        Query query = new Query();
+        query.addCriteria(Criteria.where("turma").is(nomeTurma).and("graduacao").elemMatch(Criteria.where("status").is(true)));
+        return this.mongoTemplate.exists(query, Aluno.class);
+    }
+
     private Turma encontrarTurmaPeloNome(String nomeTurma) {
         return this.turmaRepositorio.findByNome(nomeTurma).orElseThrow(
                 () -> new EntityNotFoundException("Turma com esse nome não existe"));
+    }
+
+    private Admin encontrarAdminPeloEmail(String email) {
+        return this.adminRepositorio.findByEmailAdmin(email).orElseThrow(
+                () -> new EntityNotFoundException("Admin com esse email não existe"));
     }
 }
